fix pr remarks
