+++ azure-kusto-data/azure/kusto/data/_version.py
@@ -1,2 +1,2 @@
 """Version file."""
-VERSION = "0.0.44"
+VERSION = "0.0.43"
