+++ azure-kusto-data/azure/kusto/data/exceptions.py
@@ -50,3 +50,10 @@ class KustoAuthenticationError(KustoClientError):
 
     def __repr__(self):
         return "KustoAuthenticationError('{}', '{}', '{}')".format(self.authentication_method, repr(self.exception), self.kwargs)
+
+
+class KustoAsyncRuntimeError(RuntimeError):
+    """Raised when trying to use sync flow from an async environment"""
+
+    def __init__(self, func_name):
+        super().__init__("Kusto driver has detected a running event loop, consider using {func_name}_async instead of {func_name}".format(func_name=func_name))
